Vibration feature 리팩토링
